First pass at health_stats library
This is intended to allow parsing and analyzing of a large variety
of health-related statistics.  This first commit contains the basic
framework required to load a mysugr_csv file and store it into a
data structure that can parse it into something meaningful.

Two sample reports are included, though their organization needs
to be rethought.